src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
anyio==4.0.0
    # via httpx
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
exceptiongroup==1.1.3
    # via anyio
h11==0.14.0
    # via httpcore
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==1.0.2
    # via httpx
httpx[http2]==0.25.1
    # via -r requirements.in
hyperframe==6.0.1
    # via h2
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
praw==7.7.1
    # via -r requirements.in
prawcore==2.4.0
    # via praw
python-dotenv==1.0.0
    # via -r requirements.in
requests==2.31.0
    # via
    #   prawcore
    #   update-checker
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
update-checker==0.18.0
    # via praw
urllib3==2.0.7
    # via requests
websocket-client==1.6.4
    # via praw
